+++ cpp/benchmarks/interop/interop.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, NVIDIA CORPORATION.
+ * Copyright (c) 2024-2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,13 +17,14 @@
 #include <benchmarks/common/generate_input.hpp>
 #include <benchmarks/common/table_utilities.hpp>
 
+#include <cudf_test/nanoarrow_utils.hpp>
+
 #include <cudf/interop.hpp>
 
 #include <thrust/iterator/counting_iterator.h>
 
 #include <nanoarrow/nanoarrow.hpp>
 #include <nanoarrow/nanoarrow_device.h>
-#include <nanoarrow_utils.hpp>
 #include <nvbench/nvbench.cuh>
 
 #include <algorithm>
