+++ cpp/tests/interop/from_arrow_test.cpp
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-#include <tests/interop/arrow_utils.hpp>
-#include <tests/interop/nanoarrow_utils.hpp>
+#include "arrow_utils.hpp"
 
 #include <cudf_test/base_fixture.hpp>
 #include <cudf_test/column_utilities.hpp>
 #include <cudf_test/column_wrapper.hpp>
+#include <cudf_test/nanoarrow_utils.hpp>
 #include <cudf_test/table_utilities.hpp>
 #include <cudf_test/type_lists.hpp>
